Added "are you sure?" question step in the delete command.

+++ PswManagerLibrary/Commands/DeleteCommand.cs
@@ -3,15 +3,18 @@ using PswManagerCommands.AbstractCommands;
 using PswManagerCommands.Validation;
 using PswManagerLibrary.Extensions;
 using PswManagerLibrary.Storage;
+using PswManagerLibrary.UIConnection;
 
 namespace PswManagerLibrary.Commands {
     public class DeleteCommand : BaseCommand {
 
         //todo - implement tests for this command
         private readonly IPasswordManager pswManager;
+        private readonly IUserInput userInput;
 
-        public DeleteCommand(IPasswordManager pswManager) {
+        public DeleteCommand(IPasswordManager pswManager, IUserInput userInput) {
             this.pswManager = pswManager;
+            this.userInput = userInput;
         }
 
         public override string GetSyntax() {
@@ -26,6 +29,9 @@ namespace PswManagerLibrary.Commands {
         }
 
         protected override CommandResult RunLogic(string[] arguments) {
+            var result = userInput.YesOrNo("Are you sure? This account will be deleted forever.");
+            if(result == false) { return new CommandResult("The operation has been stopped.", false); }
+
             pswManager.DeletePassword(arguments[0]);
 
             return new CommandResult("Account deleted successfully.", true);
